chipchat/requirements.txt
streamlit==1.32.0
requests==2.31.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
langchain==0.0.267
google-colab==0.0.1a1
pandas==2.0.3 